+++ ThinkingHome.Plugins.WebUi/Resources/Application/lib.js
@@ -87,7 +87,6 @@ define([
             ApplicationBlock: applicationBlock,
             AppSection: appSection,
             getJSON: getJSON,
-            loadModel: loadModel,
             serializeForm: serializeForm
         },
         ajax: {